// src/auth/dto/sign-up.dto.ts
import { z } from 'zod'

export const SignUpSchema = z.object({
  name: z.string().min(3, 'Name must be at least 3 characters'),
  email: z.string().email('Invalid email address'),
  password: z.string().min(6, 'Password must be at least 6 characters'),
})

export type SignUpDto = z.infer<typeof SignUpSchema>